+++ CHANGES.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.5] - 2024-01-14
+## [1.0.7] - 2024-01-16
+
+### Fixed
+- Spelling and docs. Make file now lints docs and runs spell check tools.
+
+
+## [1.0.6] - 2024-01-15
+
+### Fixed
+- Add "basic_test.sh" and fix issue found with it.
+
+
+## [1.0.5] - 2024-01-15
 
 ### Fixed
 - Needs packaging lib in deps